+++ src/config.ts
@@ -3,6 +3,13 @@ import * as vscode from "vscode";
 export function getAutocallPaths(): vscode.Uri[] {
   const config = vscode.workspace.getConfiguration("sas");
   const sasautos = config.get<string[]>("sasautos") || [];
+  // config.get() doesn't guarantee the result is of type T.
+  if (!Array.isArray(sasautos)) {
+    vscode.window.showWarningMessage(
+      "The `sas.sasautos` setting has a malformed value. It'll be ignored.",
+    );
+    return [];
+  }
   return sasautos.flatMap(substituteVariables).map(vscode.Uri.file);
 }